# Conflicts:
#	BARRACUDA.sh.txt
#	BOA.sh.txt
#	CHANGELOG.txt
#	aegir/conf/barracuda-release.txt
#	aegir/conf/default.boa_platform_control.ini
#	aegir/conf/default.boa_site_control.ini
#	aegir/conf/octopus-release.txt
#	aegir/conf/release.txt
#	aegir/helpers/websh.sh.txt
#	aegir/scripts/AegirSetupA.sh.txt
#	aegir/scripts/AegirSetupB.sh.txt
#	aegir/scripts/AegirSetupC.sh.txt
#	aegir/scripts/AegirSetupM.sh.txt
#	aegir/scripts/AegirUpgrade.sh.txt
#	aegir/tools/BOND.sh.txt
#	aegir/tools/bin/autobeowulf
#	aegir/tools/bin/autochimaera
#	aegir/tools/bin/autoupboa
#	aegir/tools/bin/backboa
#	aegir/tools/bin/barracuda
#	aegir/tools/bin/boa
#	aegir/tools/bin/cluster
#	aegir/tools/bin/duobackboa
#	aegir/tools/bin/mycnfup
#	aegir/tools/bin/octopus
#	aegir/tools/bin/weblogx
#	aegir/tools/system/clear.sh
#	aegir/tools/system/conf/lshell.conf
#	aegir/tools/system/daily.sh
#	aegir/tools/system/manage_ltd_users.sh
#	aegir/tools/system/manage_solr_config.sh
#	aegir/tools/system/mysql_hourly.sh
#	lib/functions/helper.sh.inc
#	lib/functions/jetty.sh.inc
#	lib/functions/master.sh.inc
#	lib/functions/php.sh.inc
#	lib/functions/satellite.sh.inc
#	lib/functions/system.sh.inc
#	lib/settings/barracuda.sh.cnf
#	lib/settings/octopus.sh.cnf

+++ aegir/tools/system/clear.sh
@@ -100,7 +100,7 @@ find_fast_mirror() {
     _USE_MIR="files.aegir.cc"
   fi
   urlDev="http://${_USE_MIR}/dev"
-  urlHmr="http://${_USE_MIR}/versions/stable/boa/aegir"
+  urlHmr="http://${_USE_MIR}/versions/dev/boa/aegir"
 }
 
 if [ ! -e "/var/run/boa_run.pid" ]; then
@@ -115,7 +115,7 @@ if [ ! -e "/var/run/boa_run.pid" ]; then
       fi
       echo "curl install" | dpkg --set-selections &> /dev/null
       apt_clean_update
-      apt-get install curl ${aptYesUnth} &> /dev/null
+      apt-get install curl ${aptYesUnth} -fu --reinstall &> /dev/null
       mkdir -p /var/backups/libcurl
       mv -f /usr/local/lib/libcurl* /var/backups/libcurl/ &> /dev/null
       mv -f /usr/local/lib/pkgconfig/libcurl* /var/backups/libcurl/ &> /dev/null
@@ -126,14 +126,19 @@ if [ ! -e "/var/run/boa_run.pid" ]; then
   rm -f /var/backups/BOA.sh.txt.hourly*
   find_fast_mirror
   curl -L -k -s \
-    --max-redirs 10 \
-    --retry 10 \
+    --max-redirs 5 \
+    --retry 5 \
     --retry-delay 5 \
-    -A iCab "http://${_USE_MIR}/versions/stable/boa/BOA.sh.txt" \
+    -A iCab "http://${_USE_MIR}/versions/dev/boa/BOA.sh.txt" \
     -o /var/backups/BOA.sh.txt.hourly
-  bash /var/backups/BOA.sh.txt.hourly
-  rm -f /var/backups/BOA.sh.txt.hourly*
-  sleep 3
+  wait
+  if [ -e "/var/backups/BOA.sh.txt.hourly" ]; then
+    bash /var/backups/BOA.sh.txt.hourly
+    wait
+    rm -f /var/backups/BOA.sh.txt.hourly*
+  else
+    echo "Not available /var/backups/BOA.sh.txt.hourly"
+  fi
   bash /opt/local/bin/autoupboa
 fi
 
@@ -145,9 +150,10 @@ if [[ "${checkVn}" =~ "===" ]] || [ -z "${checkVn}" ]; then
     checkVn="whereis barracuda_log.txt"
   fi
 fi
-crlHead="-I -k -s --retry 8 --retry-delay 8"
-urlBpth="http://${_USE_MIR}/versions/stable/boa/aegir/tools/bin"
+crlHead="-I -k -s --retry 5 --retry-delay 5"
+urlBpth="http://${_USE_MIR}/versions/dev/boa/aegir/tools/bin"
 curl ${crlHead} -A "${checkVn}" "${urlBpth}/thinkdifferent" &> /dev/null
+wait
 
 renice ${_B_NICE} -p $$ &> /dev/null
 service ssh restart &> /dev/null
