+++ aegir/tools/system/move_sql.sh
@@ -29,7 +29,7 @@ create_locks() {
 
 remove_locks() {
   echo "Removing locks..."
-  rm -f /var/run/boa_wait.pid
+  [ -e "/var/run/boa_wait.pid" ] && rm -f /var/run/boa_wait.pid
   rm -f /var/run/fmp_wait.pid
   rm -f /var/run/mysql_restart_running.pid
 }
@@ -79,23 +79,24 @@ stop_sql() {
   until [ -z "${_IS_NGINX_RUNNING}" ]; do
     _IS_NGINX_RUNNING=$(ps aux | grep '[n]ginx' | awk '{print $2}' 2>&1)
     echo "Waiting for Nginx graceful shutdown..."
-    sleep 3
+    sleep 1
   done
+  killall nginx &> /dev/null
   echo "Nginx stopped"
 
   echo "Stopping all PHP-FPM instances now..."
   _PHP_V="82 81 80 74 73 72 71 70 56 55 54 53"
   for e in ${_PHP_V}; do
-    if [ -e "/etc/init.d/php${e}-fpm" ]; then
+    if [ -e "/etc/init.d/php${e}-fpm" ] && [ -e "/opt/php${e}/bin/php" ]; then
       service php${e}-fpm force-quit &> /dev/null
     fi
   done
-  # kill -9 $(ps aux | grep '[p]hp-fpm' | awk '{print $2}')
   until [ -z "${_IS_FPM_RUNNING}" ]; do
     _IS_FPM_RUNNING=$(ps aux | grep '[p]hp-fpm' | awk '{print $2}' 2>&1)
     echo "Waiting for PHP-FPM graceful shutdown..."
-    sleep 3
+    sleep 1
   done
+  kill -9 $(ps aux | grep '[p]hp-fpm' | awk '{print $2}') &> /dev/null
   echo "PHP-FPM stopped"
 
   _IS_MYSQLD_RUNNING=$(ps aux | grep '[m]ysqld' | awk '{print $2}' 2>&1)
